+++ src/main/java/chess/dao/PiecesDao.java
@@ -1,14 +1,14 @@
 package chess.dao;
 
-import chess.domain.game.LoadedPiecesDto;
-import chess.domain.piece.type.Piece;
+import chess.domain.game.dto.LoadedPiecesInsertDto;
+import chess.domain.game.dto.LoadedPiecesSelectDto;
 import java.sql.Connection;
 import java.util.List;
 
 public interface PiecesDao {
     Connection getConnection();
 
-    LoadedPiecesDto findAll();
+    LoadedPiecesSelectDto findAll();
 
     void insertAll(List<Piece> pieces);
