+++ src/main/java/chess/domain/game/dto/LoadedPiecesSelectDto.java
@@ -1,15 +1,15 @@
-package chess.domain.game;
+package chess.domain.game.dto;
 
 import java.util.List;
 
-public class LoadedPiecesDto {
+public class LoadedPiecesSelectDto {
     private final List<Integer> files;
     private final List<Integer> ranks;
     private final List<String> sides;
     private final List<String> pieceTypes;
 
-    public LoadedPiecesDto(final List<Integer> files, final List<Integer> ranks, final List<String> sides,
-                           final List<String> pieceTypes) {
+    public LoadedPiecesSelectDto(final List<Integer> files, final List<Integer> ranks, final List<String> sides,
+                                 final List<String> pieceTypes) {
         this.files = files;
         this.ranks = ranks;
         this.sides = sides;
